requirements.txt
flask==3.0.0
numpy==1.26.3
pandas==2.1.4
tensorflow==2.16.1
pillow==10.1.0
scikit-learn==1.3.2
h5py==3.10.0
flask-wtf==1.2.1
flask-limiter==3.5.0
flask-swagger-ui==4.11.1
pytest==7.4.3
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0 
